+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - added additional search options to List Tasks tool
 - added option to autostart AiderDesk server on run command
 - correctly calculating cached tokens for Requesty provider models
+- added readonly server mode
+- better auto scrolling control while streaming
+- added gpt-live-transcribe model to the OpenAI voice provider
+- added option to toggle extension UI components in readonly server mode
 
 ## [0.76.0]